+++ CHANGES.rst
@@ -8,7 +8,7 @@ Version 1.1.1 (in development)
 
 Version 1.1.0 (2017.05.09)
 ==========================
-- update email address to @measurement.govt.nz
+- update email address to "measurement"
 - previous release date (in CHANGES.rst) was yyyy.dd.mm should have been yyyy.mm.dd
 - previous release should have incremented the minor number (new **update** feature)