+++ CHANGES.rst
@@ -2,8 +2,8 @@
 Release Notes
 =============
 
-Version 1.1.1 (in development)
-==============================
+Version 1.2.0 (2017.08.10)
+==========================
 - add the ``--all`` flag for the CLI
 - include ``--process-dependency-links`` argument for ``pip install``
 - create **upgrade** alias for **update**